Add files via upload
These F1 configurations can be used with the other content to enable WLED lights.  It relies currently on replacing the existing WLED chase effect with a custom one that has 3 colors (needed for F1).  Once the effect is flashed, Homeassistant can be used to send colors, positions, etc. to WLED.  A REST command added to config.yaml, a script populates variables.  I will look into sending the "new" chase effect to WLED for review, but for now this works for me.